requirements.txt
Django==5.0.2
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2
django-storages==1.14.3
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langsmith==0.1.51
openai==1.23.6
supabase==2.4.3
postgrest==0.16.3
psycopg==3.1.18
psycopg2-binary==2.9.9
pillow==9.5.0
azure-core==1.30.2
azure-storage-blob==12.20.0
isodate==0.6.1
gunicorn==21.2.0
gevent==24.2.1
qdrant-client==1.10.1